Dodana privilegija, dodani testovi, refaktoring prenesen
