--- autorizacija/src/main/java/com/example/autorizacija/AutorizacijaApplication.java
@@ -1,13 +1,0 @@
-package com.example.autorizacija;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class AutorizacijaApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(AutorizacijaApplication.class, args);
-    }
-
-}
