--- autorizacija/src/main/java/com/example/autorizacija/ImenaUloga.java
@@ -1,5 +1,0 @@
-package com.example.autorizacija;
-
-public enum ImenaUloga {
-    ADMIN, STUDENT, PROFESOR, ASISTENT, STUDENTSKA_SLUZBA
-}
