--- autorizacija/src/main/java/com/example/autorizacija/KorisnikRepozitorij.java
@@ -1,8 +1,0 @@
-package com.example.autorizacija;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface KorisnikRepozitorij extends JpaRepository<Korisnik, Long> {
-}
