--- autorizacija/src/main/java/com/example/autorizacija/PrivilegijaRepozitorij.java
@@ -1,8 +1,0 @@
-package com.example.autorizacija;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface PrivilegijaRepozitorij extends JpaRepository<Privilegija, Long> {
-}
