--- autorizacija/src/main/java/com/example/autorizacija/UlogaKontroler.java
@@ -1,7 +1,0 @@
-package com.example.autorizacija;
-
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class UlogaKontroler {
-}
