--- autorizacija/src/main/java/com/example/autorizacija/UlogaRepozitorij.java
@@ -1,9 +1,0 @@
-package com.example.autorizacija;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface UlogaRepozitorij extends JpaRepository<Uloga, Long> {
-    Uloga findBynazivUloge (ImenaUloga naziv);
-}
