--- autorizacija/src/main/resources/application.properties
@@ -1,12 +1,0 @@
-spring.jpa.hibernate.ddl-auto=none
-#spring.jpa.generate-ddl=true
-spring.jpa.show-sql=true
-spring.jpa.hibernate.naming.physical-strategy=org.hibernate.boot.model.naming.PhysicalNamingStrategyStandardImpl
-
-
-spring.datasource.url=jdbc:mysql://localhost:3306/si_autorizacija?useUnicode=true&useJDBCCompliantTimezoneShift=true&useLegacyDatetimeCode=false&serverTimezone=UTC
-spring.datasource.username=root
-#spring.datasource.password=BLysSj9ZrP
-spring.datasource.driver-class-name=com.mysql.jdbc.Driver
-
-server.port=8080
