--- autorizacija/src/test/java/com/example/autorizacija/AutorizacijaApplicationTests.java
@@ -1,16 +1,0 @@
-package com.example.autorizacija;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
-
-@RunWith(SpringRunner.class)
-@SpringBootTest
-public class AutorizacijaApplicationTests {
-
-    @Test
-    public void contextLoads() {
-    }
-
-}
