+++ demo/src/main/java/com/example/demo/ImenaUloga.java
@@ -1,0 +1,5 @@
+package com.example.demo;
+
+public enum ImenaUloga {
+    ADMIN, STUDENT, PROFESOR, ASISTENT, STUDENTSKA_SLUZBA
+}
