+++ demo/src/main/java/com/example/demo/KorisnikRepozitorij.java
@@ -1,0 +1,8 @@
+package com.example.demo;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface KorisnikRepozitorij extends JpaRepository<Korisnik, Long> {
+}
