+++ demo/src/main/java/com/example/demo/PrivilegijaKontroler.java
@@ -1,0 +1,7 @@
+package com.example.demo;
+
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class PrivilegijaKontroler {
+}
