+++ demo/src/main/java/com/example/demo/PrivilegijaRepozitorij.java
@@ -1,0 +1,9 @@
+package com.example.demo;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface PrivilegijaRepozitorij extends JpaRepository<Privilegija, Long> {
+    boolean existsBynazivPrivilegije(String privilegija);
+}
