+++ demo/src/main/resources/application.properties
@@ -1,0 +1,13 @@
+spring.jpa.hibernate.ddl-auto = none
+# spring.jpa.generate-ddl = true
+spring.jpa.show-sql = true
+spring.jpa.hibernate.naming.physical-strategy = org.hibernate.boot.model.naming.PhysicalNamingStrategyStandardImpl
+
+# Baza podataka
+# spring.datasource.url = jdbc:mysql://remotemysql.com:3306/TYQcLL35gV?useSSL=false
+spring.datasource.url=jdbc:mysql://remotemysql.com:3306/TYQcLL35gV?useUnicode=true&useJDBCCompliantTimezoneShift=true&useLegacyDatetimeCode=false&serverTimezone=UTC
+spring.datasource.username = TYQcLL35gV
+spring.datasource.password = BLysSj9ZrP
+spring.datasource.driver-class-name=com.mysql.jdbc.Driver
+
+server.port = 31915
